# CoreLoanManagement.WebApi/Models/CustomerItem.cs
using System;
using System.Collections.Generic;
using System.ComponentModel.DataAnnotations;
using System.Linq;
using System.Text;
using System.Threading.Tasks;

namespace CoreLoanManagement.WebApi.Models
{

    public class CustomerItem
    {
        public CustomerItem() { }
        public int CustomerId { get; set; }
        
        public string CustomerName { get; set; } 
        
    }
}