# CoreLoanManagement.WebSite/OpenAPIs/swagger1.json
{
  "openapi": "3.0.1",
  "info": {
    "title": "CoreLoanManagement.WebApi",
    "version": "1.0"
  },
  "paths": {
    "/CustomerItem": {
      "get": {
        "tags": [
          "CustomerItem"
        ],
        "operationId": "GetCustomerItem",
        "parameters": [
          {
            "name": "TotalCount",
            "in": "query",
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          },
          {
            "name": "PageSize",
            "in": "query",
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          },
          {
            "name": "CurrentPage",
            "in": "query",
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          },
          {
            "name": "TotalPages",
            "in": "query",
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          },
          {
            "name": "HasNext",
            "in": "query",
            "schema": {
              "type": "boolean"
            }
          },
          {
            "name": "HasPrevious",
            "in": "query",
            "schema": {
              "type": "boolean"
            }
          },
          {
            "name": "SearchKeyword",
            "in": "query",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "text/plain": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/CustomerItem"
                  }
                }
              },
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/CustomerItem"
                  }
                }
              },
              "text/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/CustomerItem"
                  }
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "CustomerItem"
        ],
        "operationId": "PostCustomerItem",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CustomerItem"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/CustomerItem"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/CustomerItem"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/CustomerItem"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CustomerItem"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/CustomerItem"
                }
              }
            }
          }
        }
      },
      "put": {
        "tags": [
          "CustomerItem"
        ],
        "operationId": "PutCustomerItem",
        "parameters": [
          {
            "name": "id",
            "in": "query",
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CustomerItem"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/CustomerItem"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/CustomerItem"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/CustomerItem"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CustomerItem"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/CustomerItem"
                }
              }
            }
          }
        }
      }
    },
    "/LoanInstallment": {
      "get": {
        "tags": [
          "LoanInstallment"
        ],
        "operationId": "PostLoanInstallment",
        "parameters": [
          {
            "name": "customerId",
            "in": "query",
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          },
          {
            "name": "pageSize",
            "in": "query",
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          },
          {
            "name": "lastId",
            "in": "query",
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "text/plain": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/CustomerLoanInstallmentDBOutItem"
                  }
                }
              },
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/CustomerLoanInstallmentDBOutItem"
                  }
                }
              },
              "text/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/CustomerLoanInstallmentDBOutItem"
                  }
                }
              }
            }
          }
        }
      }
    },
    "/LoanManager": {
      "get": {
        "tags": [
          "LoanManager"
        ],
        "operationId": "GetLoanManager",
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "text/plain": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/Customer"
                  }
                }
              },
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/Customer"
                  }
                }
              },
              "text/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/Customer"
                  }
                }
              }
            }
          }
        }
      }
    },
    "/WeatherForecast": {
      "get": {
        "tags": [
          "WeatherForecast"
        ],
        "operationId": "GetWeatherForecast",
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "text/plain": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/WeatherForecast"
                  }
                }
              },
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/WeatherForecast"
                  }
                }
              },
              "text/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/WeatherForecast"
                  }
                }
              }
            }
          }
        }
      }
    }
  },
  "components": {
    "schemas": {
      "Customer": {
        "type": "object",
        "properties": {
          "customerId": {
            "type": "integer",
            "format": "int32"
          },
          "customerName": {
            "maxLength": 50,
            "type": "string",
            "nullable": true
          },
          "loanList": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/Loan"
            },
            "nullable": true
          }
        },
        "additionalProperties": false
      },
      "CustomerItem": {
        "type": "object",
        "properties": {
          "customerId": {
            "type": "integer",
            "format": "int32"
          },
          "customerName": {
            "type": "string",
            "nullable": true
          }
        },
        "additionalProperties": false
      },
      "CustomerLoanInstallmentDBOutItem": {
        "type": "object",
        "properties": {
          "customerId": {
            "type": "integer",
            "format": "int32"
          },
          "customerName": {
            "type": "string",
            "nullable": true
          },
          "loanDescription": {
            "type": "string",
            "nullable": true
          },
          "installmentValue": {
            "type": "number",
            "format": "double"
          },
          "installmentId": {
            "type": "integer",
            "format": "int32"
          }
        },
        "additionalProperties": false
      },
      "Loan": {
        "type": "object",
        "properties": {
          "loanId": {
            "type": "integer",
            "format": "int32"
          },
          "loanDescription": {
            "type": "string",
            "nullable": true
          },
          "loanValue": {
            "type": "number",
            "format": "double"
          },
          "customer": {
            "$ref": "#/components/schemas/Customer"
          }
        },
        "additionalProperties": false
      },
      "WeatherForecast": {
        "type": "object",
        "properties": {
          "date": {
            "type": "string",
            "format": "date-time"
          },
          "temperatureC": {
            "type": "integer",
            "format": "int32"
          },
          "temperatureF": {
            "type": "integer",
            "format": "int32",
            "readOnly": true
          },
          "summary": {
            "type": "string",
            "nullable": true
          }
        },
        "additionalProperties": false
      }
    }
  }
}